pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

#[tool.setuptools.packages.find]
#where = ["src"]  # list of folders that contain the packages (["."] by default)

[project]
name = "seam-carving"
dynamic = ["version"]
dependencies = ["numpy", "pillow"]
requires-python = ">=3.8"
authors = [{ name = "Student Name", email = "student@example.com" }]
description = "Seam carving algorithm for content-aware image resizing"
readme = "README.md"
license = { file = "LICENSE" }
keywords = [
  "seam-carving",
  "image-processing",
  "image-resizing",
  "content-aware",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
]

[project.optional-dependencies]
dev = ["mypy", "ruff", "pytest", "pytest-cov", "tox", "pre-commit"]

[tool.setuptools.package-data]
seam_carving = ["py.typed"]

[project.urls]
Repository = "https://github.com/student/seam-carving.git"

###################################################################################################
[tool.pytest.ini_options]
addopts = "--cov=seam_carving"
testpaths = ["tests"]

[tool.mypy]
mypy_path = "src"
check_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
no_implicit_reexport = true

[tool.ruff]
src = ["src"]

[tool.ruff.lint]
extend-select = ["UP", "I"]
